revert changes made for bsd compatiblity
the the spl layer function kmem_cache_alloc does not support the KM_ZERO
flag unlike the standard allocation function which is a difference from
freebsd which does support this. initially i implemented a zalloc
variant to fix this, unfortunatly this raised a ASSERT since KM_ZERO is
masked out in several checks. so i decided to completelly remove this
change and implemented a single zero memset on the one and only
occurence of such a requirement

Signed-off-by: Sebastian Gottschall <s.gottschall@dd-wrt.com>