vmem allocation pool (experimental)
this new allocator guarantees that there is always
memory available for a compression (and more important)
decompression task. if the system is running out of memory,
decompression may fail and causes data corruption.
we deal with that situation by allocating a static memory
pool which is used as fallback if allocation fails.
this pool will be secured by a barrier and does not allow
threading anymore, but ensures that the data will never
be corrupted.

this new code still contains some debug output which will
be removed later

Signed-off-by: Sebastian Gottschall <s.gottschall@dd-wrt.com>