remove debug code and change allocation scheme
standard compression context will now be allocated using KM_NOSLEEP
but decompression is forced to use KM_SLEEP to ensure that decompression
never fails. however. if allocation fails, there is still a fallback
allocator in place which is often used by the compression workspace
since KM_NOSLEEP fails in 70% of all allocation tries of bigger blocks
on Linux. So this might be a expensive way of allocating memory.
I will do some more benchmarks on this topic. if it shows a performance
decrease comparing with KM_SLEEP i will look for a different approach.

Signed-off-by: Sebastian Gottschall <s.gottschall@dd-wrt.com>